pyproject.toml
[tool.poetry]
name = "oeh-search-meta"
version = "0.1.0"
description = ""
authors = [
    "Robert Meissner <robert.meissner@codecentric.de>",
    "Martin Rückl <martin.rueckl@codecentric.de>",
]
packages = [
    { include = "metalookup", from = "src" },
]

[tool.poetry.scripts]
meta-lookup = 'metalookup.main:main'

[tool.poetry.dependencies]
python = "^3.9.1"
requests = "^2.27"
lxml = "^4.9"
tldextract = "^3.3"
"pdfminer.six" = "^20201018"
PyPDF2 = "^1.28"
aiohttp = "^3.8"
beautifulsoup4 = "^4.11"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"
databases = {extras = ["sqlite", "postgresql"], version = "^0.6.0"}
adblockparser = "^0.7"
google-re2 = "^0.2.20220601"
playwright = "^1.24.0"

[tool.poetry.dev-dependencies]
pylint = "^2.14"
pre-commit = "^2.19"
pytest = "^6.2"
pytest-cov = "^2.12"
pytest-mock = "^3.7"
mkdocs = "^1.3"
locust = "^2.9"
httpx = "^0.23.0"
pytest-asyncio = "^0.18.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'

[tool.isort]
multi_line_output = 3
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"
line_length = 120

[tool.pytest.ini_options]
minversion = "6.0"
asyncio_mode = "auto"
log_cli = "true"
log_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s %(name)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = [
    "tests",
]
